bounty_platform/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1
celery==5.3.6
fastapi-limiter==0.1.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
openpyxl==3.1.2
pillow==10.2.0
captcha==0.5.0
psutil==5.9.8
requests==2.31.0